% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDARSMerge.R
\name{CDARSMerge}
\alias{CDARSMerge}
\title{Quick merge CDARS files}
\usage{
CDARSMerge(class, to_csv = FALSE)
}
\arguments{
\item{class}{classes of files to be combined}

\item{to_csv}{convert to csv if TRUE, or to SPSS}
}
\description{
Merge CDARS raw xlsx files of the same types (column names) under the same folder automatically, save to SPSS format
}
\details{
supports xlsx only
}
\examples{
\dontrun{
# CDARSMerge(class = c("Dx","Px"))
# CDARSMerge(class = c("alt","DeAth"),TRUE)
}

}
\seealso{
\code{\link[readxl]{read_xlsx}}

\code{\link[readxl]{excel_sheets}}
}
