% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1Par.R
\name{Table1Par}
\alias{Table1Par}
\title{Select Parameters for Table1}
\usage{
Table1Par()
}
\description{
Select Parameters for Table1
}
\details{
supports xlsx or spss files only
}
\examples{
\dontrun{
# Table1Par()
}
}
\seealso{
\code{\link[readxl]{read_xlsx}}

\code{\link[haven]{read_sav}}
}
